API

XCB.keysym_names_translation

WindowAbstractions.KeySymbol

WindowAbstractions.KeySymbol

XCB.Connection

XCB.Connection

XCB.ConnectionError

XCB.FlushError

XCB.GraphicsContext

XCB.Handle

XCB.RequestError

XCB.Setup

XCB.XCBWindow

Base.run

XCB.check

XCB.check_flush

XCB.check_request

XCB.keysym_name_to_keysymbol

XCB.@check

XCB.@flush

XCB.keysym_names_translation — Constant

Translation Dict from a keysym obtained via XCB to a standard key name defined in WindowAbstractions.

WindowAbstractions.KeySymbol — Method

Get a KeySymbol from a physical key name.

KeySymbol(km::XCB.Keymap, key_name::Any) -> WindowAbstractions.KeySymbol

WindowAbstractions.KeySymbol — Method

Get a KeySymbol from a keycode.

KeySymbol(km::XCB.Keymap, keycode::Integer) -> WindowAbstractions.KeySymbol

XCB.Connection — Type

Connection to the X server.

mutable struct Connection <: XCB.Handle

h::Any

Opaque handle to the connection, used for API calls.

XCB.Connection — Method

Initialize a connection to the X server.

Connection(; display) -> Connection

XCB.ConnectionError — Type

The X Server reported a connection error.

struct ConnectionError <: Exception

msg::Any

code::Any

XCB.FlushError — Type

Error when flushing a connection to the X Server.

struct FlushError <: Exception

code::Any

XCB.GraphicsContext — Type

Graphics context attached to a window. Used to register drawing commands on the window surface.

mutable struct GraphicsContext

conn::Any

id::Any

window::Any

mask::Any

value_list::Any

XCB.Handle — Type

Structures which contain a handle (opaque pointer) as primary data. Those structures are usually defined as mutable so that a finalizer can be registered. Any Handle structure is automatically converted to its handle data on ccalls, through unsafe_convert.

abstract type Handle

XCB.RequestError — Type

A request to the X Server returned with an error.

struct RequestError <: Exception

msg::Any

XCB.Setup — Type

Connection setup handle and data.

struct Setup <: XCB.Handle

h::Any

Handle to the setup, used for API calls.

value::XCB.Libxcb.xcb_setup_t

Setup value, obtained when dereferencing its handle.

XCB.XCBWindow — Type

Window type used with the XCB API.

mutable struct XCBWindow <: WindowAbstractions.AbstractWindow

conn::Connection

id::Any

parent_id::Any

visual_id::Any

class::Any

depth::Any

mask::Any

value_list::Any

ctx::Any

delete_request::Any

Base.run — Method

Run an EventLoop attached to a XWindowHandler instance.

run(event_loop::WindowAbstractions.EventLoop{XWindowHandler}, ::WindowAbstractions.Synchronous; warn_unknown, poll, kwargs...)

XCB.check — Method

Check that the connection to the X server was successful. Throws a ConnectionError if the connection failed.

check(connection::Connection) -> Connection

XCB.check_flush — Method

Check that the flush was successful, throwing a FlushError if the code is negative.

check_flush(code::Any)

XCB.check_request — Method

Check that the request was successfully handled by the server, throwing a RequestError if the request failed.

check_request(conn::Any, request::Any; raise)

XCB.keysym_name_to_keysymbol — Method

Get a KeySymbol from a keysym name.

keysym_name_to_keysymbol(keysym_name::Any) -> WindowAbstractions.KeySymbol

XCB.@check — Macro

Check the value returned by the function call request with check_request.

Wraps request with check_request. The Connection argument is taken as the first argument of the function call expression request. The request is transformed to be checkable, through the functions xcb*checked (or xcb* if there exists a xcb*_unchecked version). If no checkable substitute is found, an ArgumentError is raised.

TODO: @macroexpand example

XCB.@flush — Macro

Flush a connection attached to a request expr.

The connection is taken to be the first argument of expr. expr can be a call to XCB.@check.

Examples

julia> @macroexpand @flush xcb_unmap_window(win.conn, win.id)
quote
    xcb_unmap_window(win.conn, win.id)
    (flush)(win.conn)
end