API

XCB.Connection

XCB.Connection

XCB.ConnectionError

XCB.FlushError

XCB.GraphicsContext

XCB.Handle

XCB.RequestError

XCB.Setup

XCB.XCBWindow

XCB.XCBWindow

XCB.XCBWindow

XCB.check

XCB.check_flush

XCB.check_request

XCB.response_type_xcb

XCB.state_xcb

XCB.@check

XCB.@flush

XCB.Connection — Type

Connection to the X server.

mutable struct Connection <: XCB.Handle

h::Ptr{Nothing}: Opaque handle to the connection, used for API calls.

XCB.Connection — Method

Initialize a connection to the X server.

Connection(; display) -> Connection

XCB.ConnectionError — Type

The X Server reported a connection error.

struct ConnectionError <: Exception

msg::Any

code::Any

XCB.FlushError — Type

Error when flushing a connection to the X Server.

struct FlushError <: Exception

code::Any

XCB.GraphicsContext — Type

Graphics context attached to a window. Used to register drawing commands on the window surface.

mutable struct GraphicsContext

conn::Connection

id::UInt32

XCB.Handle — Type

Structures which contain a handle (opaque pointer) as primary data. Those structures are usually defined as mutable so that a finalizer can be registered. Any Handle structure is automatically converted to its handle data on ccalls, through unsafe_convert.

abstract type Handle

XCB.RequestError — Type

A request to the X Server returned with an error.

struct RequestError <: Exception

msg::Any

XCB.Setup — Type

Connection setup handle and data.

struct Setup <: XCB.Handle

h::Ptr{xcb_setup_t}: Handle to the setup, used for API calls.

value::xcb_setup_t: Setup value, obtained when dereferencing its handle.

XCB.XCBWindow — Type

Window type used with the XCB API.

mutable struct XCBWindow <: AbstractWindow

conn::Connection

id::UInt32

screen::xcb_screen_t

frame_extents::Union{Nothing, NTuple{4, Int64}}

gc::Union{Nothing, GraphicsContext}

position::Tuple{Int64, Int64}

extent::Tuple{Int64, Int64}

ignore_next_moves::Int64

atoms::Dict{Symbol, UInt32}

XCB.XCBWindow — Method

Create a new window on the provided screen.

XCBWindow(
    conn,
    screen::xcb_screen_t;
    depth,
    offset,
    width,
    height,
    border_width,
    class,
    window_title,
    icon_title,
    map,
    attributes,
    values,
    with_decorations
) -> XCBWindow

XCB.XCBWindow — Method

Create a new window on the current screen.

XCBWindow(conn::Connection; kwargs...) -> XCBWindow

XCB.check — Method

Check that the connection to the X server was successful. Throws a ConnectionError if the connection failed.

check(connection::Connection) -> Connection

XCB.check_flush — Method

Check that the flush was successful, throwing a FlushError if the code is negative.

check_flush(code)

XCB.check_request — Method

Check that the request was successfully handled by the server, throwing a RequestError if the request failed.

The severity level can be set via level with one of the two following values:

- `:warn`: logs a warning with `@warn` (default).
- `:error`: raise a `RequestError`.

check_request(conn, request; level)

XCB.response_type_xcb — Method

Translate an action into its corresponding XCB value.

response_type_xcb(event::Event) -> Int64

XCB.state_xcb — Method

Translate mouse or modifier state into the corresponding XCB value.

state_xcb(event::Event) -> Any

XCB.@check — Macro

Check the value returned by the function call request with check_request.

The severity level (:error or :warn) can be supplied as first argument. By default, the severity level is :warn.

Wraps request with check_request. The Connection argument is taken as the first argument of the function call expression request. The request is transformed to be checkable, through the functions xcb*checked (or xcb* if there exists a xcb*_unchecked version). If no checkable substitute is found, an ArgumentError is raised.

TODO: @macroexpand example

XCB.@flush — Macro

Flush a connection attached to a request expr.

The connection is taken to be the first argument of expr. expr can be a call to XCB.@check.

Examples

julia> @macroexpand @flush xcb_unmap_window(window.conn, window.id)
quote
    xcb_unmap_window(window.conn, window.id)
    (flush)(window.conn)
end