API

WindowAbstractions.KeySymbol

WindowAbstractions.KeySymbol

XCB.Connection

XCB.Connection

XCB.ConnectionError

XCB.FlushError

XCB.GraphicsContext

XCB.Handle

XCB.Keymap

XCB.Keymap

XCB.RequestError

XCB.Setup

XCB.XCBWindow

XCB.XCBWindow

XCB.XCBWindow

XCB.XCBWindow

Base.run

XCB.check

XCB.check_flush

XCB.check_request

XCB.key_event_from_name

XCB.key_name

XCB.key_symbol_from_keysym_string

XCB.keycode

XCB.keycode

XCB.keystroke_info

XCB.keysym_string

XCB.response_type_xcb

XCB.standardize_key_symbol

XCB.state_xcb

XCB.@check

XCB.@flush

WindowAbstractions.KeySymbol — Method

Generate a KeySymbol from a keycode.

KeySymbol(km::XCB.Keymap, keycode::Integer) -> KeySymbol

WindowAbstractions.KeySymbol — Method

Generate a KeySymbol from a key name.

KeySymbol(km::XCB.Keymap, key_name::Symbol) -> KeySymbol

XCB.Connection — Type

Connection to the X server.

mutable struct Connection <: XCB.Handle

h::Any

Opaque handle to the connection, used for API calls.

XCB.Connection — Method

Initialize a connection to the X server.

Connection(; display) -> Connection

XCB.ConnectionError — Type

The X Server reported a connection error.

struct ConnectionError <: Exception

msg::Any

code::Any

XCB.FlushError — Type

Error when flushing a connection to the X Server.

struct FlushError <: Exception

code::Any

XCB.GraphicsContext — Type

Graphics context attached to a window. Used to register drawing commands on the window surface.

mutable struct GraphicsContext

conn::Connection

id::UInt32

XCB.Handle — Type

Structures which contain a handle (opaque pointer) as primary data. Those structures are usually defined as mutable so that a finalizer can be registered. Any Handle structure is automatically converted to its handle data on ccalls, through unsafe_convert.

abstract type Handle

XCB.Keymap — Type

Keymap used to encode information regarding keyboard layout, and country and language codes.

A string representation can be obtained from a Keymap by using String(keymap).

mutable struct Keymap <: XCB.Handle

h::Ptr{Nothing}

ctx::Ptr{Nothing}

state::Ptr{Nothing}

conn::Connection

device_id::Any

XCB.Keymap — Method

Construct a Keymap using conn as the connection to the X Server.

It uses XKB, the X Keyboard extension, which must be initialized for each connection, typically when creating a keymap (it is unlikely to be used before then). If this is not your first call to this constructor with this connection, you should set setup_xkb to false.

Keymap(conn::Connection; setup_xkb) -> XCB.Keymap

XCB.RequestError — Type

A request to the X Server returned with an error.

struct RequestError <: Exception

msg::Any

XCB.Setup — Type

Connection setup handle and data.

struct Setup <: XCB.Handle

h::Any

Handle to the setup, used for API calls.

value::XCB.Libxcb.xcb_setup_t

Setup value, obtained when dereferencing its handle.

XCB.XCBWindow — Type

Window type used with the XCB API.

mutable struct XCBWindow <: AbstractWindow

conn::Connection

id::UInt32

parent_id::UInt32

visual_id::UInt32

delete_request::UInt32

gc::Union{Nothing, GraphicsContext}

XCB.XCBWindow — Method

Create a new window whose parent is given by parent_id and visual by visual_id.

XCBWindow(conn, parent_id, visual_id; depth, x, y, width, height, border_width, class, window_title, icon_title, map, attributes, values) -> XCBWindow

XCB.XCBWindow — Method

Create a new window on the provided screen.

XCBWindow(conn::Connection, screen; kwargs...) -> XCBWindow

XCB.XCBWindow — Method

Create a new window on the current screen.

XCBWindow(conn::Connection; kwargs...) -> XCBWindow

Base.run — Function

Run an EventLoop attached to a XWindowManager instance.

run(wm::XWindowManager, ::Synchronous)
run(wm::XWindowManager, ::Synchronous, execute_callback; poll, kwargs...)

XCB.check — Method

Check that the connection to the X server was successful. Throws a ConnectionError if the connection failed.

check(connection::Connection) -> Connection

XCB.check_flush — Method

Check that the flush was successful, throwing a FlushError if the code is negative.

check_flush(code)

XCB.check_request — Method

Check that the request was successfully handled by the server, throwing a RequestError if the request failed.

The severity level can be set via level with one of the two following values:

- `:warn`: logs a warning with `@warn` (default).
- `:error`: raise a `RequestError`.

check_request(conn, request; level)

XCB.key_event_from_name — Method

Produce a key event based on a key name, a modifier state and an action using a keymap.

key_event_from_name(km::XCB.Keymap, key_name::Symbol, modifiers::KeyModifierState, action::KeyAction) -> KeyEvent

XCB.key_name — Method

Obtain the name of a physical key identified by its keycode using a keymap.

key_name(km::XCB.Keymap, keycode::Integer) -> Symbol

XCB.key_symbol_from_keysym_string — Method

Generate a KeySymbol from a XCB keysym string.

key_symbol_from_keysym_string(keysym::AbstractString) -> KeySymbol

XCB.keycode — Method

Obtain a keycode from a key name using a keymap.

keycode(km::XCB.Keymap, name::Symbol) -> UInt32

XCB.keycode — Method

Extract a keycode from an key event.

keycode(wm::XWindowManager, details::EventDetails{var"#s3784", W, F} where {var"#s3784"<:KeyEvent, W<:AbstractWindow, F<:AbstractFloat}) -> UInt32

XCB.keystroke_info — Method

Return condensed information regarding a keystroke as a String. Includes key name, keycode, character input and key symbol.

keystroke_info(wm::XWindowManager, event_details::EventDetails) -> String

XCB.keysym_string — Method

Return a String representation of an integer XCB keysym.

keysym_string(keysym::Integer; max_chars) -> String

XCB.response_type_xcb — Function

Translate an action into its corresponding XCB value.

XCB.standardize_key_symbol — Method

Transform a key symbol obtained via XCB to a standard symbol defined in WindowAbstractions.

standardize_key_symbol(key_symbol::Symbol) -> Symbol

XCB.state_xcb — Function

Translate mouse or modifier state into the corresponding XCB value.

XCB.@check — Macro

Check the value returned by the function call request with check_request.

The severity level (:error or :warn) can be supplied as first argument. By default, the severity level is :warn.

Wraps request with check_request. The Connection argument is taken as the first argument of the function call expression request. The request is transformed to be checkable, through the functions xcb*checked (or xcb* if there exists a xcb*_unchecked version). If no checkable substitute is found, an ArgumentError is raised.

TODO: @macroexpand example

XCB.@flush — Macro

Flush a connection attached to a request expr.

The connection is taken to be the first argument of expr. expr can be a call to XCB.@check.

Examples

julia> @macroexpand @flush xcb_unmap_window(win.conn, win.id)
quote
    xcb_unmap_window(win.conn, win.id)
    (flush)(win.conn)
end